fixed seo issue
